Review: move everything to diffusion-cli for now
